requirements.txt
albumentations==1.2.1
axial_positional_embedding==0.2.1
blobfile==1.3.4
einops==0.3.0
ftfy==6.1.1
imageio==2.9.0
linear_attention_transformer==0.19.1
matplotlib==3.4.3
mpi4py==3.1.3
numpy==1.21.2
omegaconf==2.1.1
opencv_python_headless==4.6.0.66
Pillow==9.2.0
prettytable==3.3.0
pytorch_fid==0.1.1
pytorch_lightning==1.4.2
regex==2022.7.9
requests==2.25.1
scikit_learn==1.1.2
tokenizers==0.13.0
torch==1.9.0+cu111
torchvision==0.10.0+cu111
tqdm==4.61.2
wandb==0.12.0
